requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
baron==0.9
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.7
django-allauth==0.45.0
django-cache-url==3.2.3
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-environs==0.1.1
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
environs==9.3.3
gunicorn==20.1.0
idna==3.2
marshmallow==3.13.0
multidict==5.1.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
redbaron==0.9.2
requests==2.26.0
requests-oauthlib==1.3.0
rply==0.7.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
yarl==1.6.3
